graphing code

+++ lib_bgp_data/rovpp/tables.py
@@ -363,31 +363,31 @@ class Simulation_Results_Agg_Table(Generic_Table):
         
                 --adopting traceback
                 trace_hijacked_adopting::decimal / trace_total_adopting::decimal AS trace_hijacked_adopting,
-                (trace_blackholed_adopting + no_rib_adopting)::decimal / trace_total_adopting::decimal AS trace_disconnected_adopting,
+                (trace_blackholed_adopting::decimal + no_rib_adopting::decimal)::decimal / trace_total_adopting::decimal AS trace_disconnected_adopting,
                 trace_nothijacked_adopting::decimal / trace_total_adopting::decimal AS trace_connected_adopting,
         
                 --collateral traceback
                 trace_hijacked_collateral::decimal / trace_total_collateral::decimal AS trace_hijacked_collateral,
-                (trace_blackholed_collateral + no_rib_collateral)::decimal / trace_total_collateral::decimal AS trace_disconnected_collateral,
+                (trace_blackholed_collateral::decimal + no_rib_collateral::decimal)::decimal / trace_total_collateral::decimal AS trace_disconnected_collateral,
                 trace_nothijacked_collateral::decimal / trace_total_collateral::decimal AS trace_connected_collateral,
         
                 --adopting control plane
                 c_plane_has_attacker_prefix_origin_adopting::decimal / trace_total_adopting::decimal AS c_plane_hijacked_adopting,
-                (c_plane_has_bhole_adopting + no_rib_adopting)::decimal / trace_total_adopting::decimal AS c_plane_disconnected_adopting,
+                (c_plane_has_bhole_adopting::decimal + no_rib_adopting::decimal)::decimal / trace_total_adopting::decimal AS c_plane_disconnected_adopting,
                 c_plane_has_only_victim_prefix_origin_adopting::decimal / trace_total_adopting::decimal AS c_plane_connected_adopting,
         
                 --collateral control plane
                 c_plane_has_attacker_prefix_origin_collateral::decimal / trace_total_collateral::decimal AS c_plane_hijacked_collateral,
-                (c_plane_has_bhole_collateral + no_rib_collateral)::decimal / trace_total_collateral::decimal AS c_plane_disconnected_collateral,
+                (c_plane_has_bhole_collateral::decimal + no_rib_collateral::decimal)::decimal / trace_total_collateral::decimal AS c_plane_disconnected_collateral,
                 c_plane_has_only_victim_prefix_origin_collateral::decimal / trace_total_collateral::decimal AS c_plane_connected_collateral,
         
                 --adopting hidden hijacks
                 visible_hijacks_adopting::decimal / trace_total_adopting::decimal AS visible_hijacks_adopting,
-                (trace_hijacked_adopting - visible_hijacks_adopting)::decimal / trace_total_adopting::decimal AS hidden_hijacks_adopting,
+                (trace_hijacked_adopting::decimal - visible_hijacks_adopting::decimal)::decimal / trace_total_adopting::decimal AS hidden_hijacks_adopting,
         
                 --collateral hidden hijacks
                 visible_hijacks_collateral::decimal / trace_total_collateral::decimal AS visible_hijacks_collateral,
-                (trace_hijacked_collateral - visible_hijacks_collateral)::decimal / trace_total_collateral::decimal AS hidden_hijacks_collateral
+                (trace_hijacked_collateral::decimal - visible_hijacks_collateral::decimal)::decimal / trace_total_collateral::decimal AS hidden_hijacks_collateral
         
             FROM {Simulation_Results_Table.name}
         );"""
